check for existence of GA data before sending pageviews
