+++ app/client/user.ts
@@ -1,6 +1,5 @@
 import Raven from "raven-js";
 import ReactDOM from "react-dom";
-import { trackPath } from "./analytics";
 import { BrowserUser } from "./components/user";
 
 declare var WEBPACK_BUILD: string;
@@ -13,4 +12,4 @@ if (typeof window !== "undefined" && window.guardian && window.guardian.dsn) {
 }
 
 const element = document.getElementById("app");
-ReactDOM.hydrate(BrowserUser(trackPath), element);
+ReactDOM.hydrate(BrowserUser, element);
